Create CardOrderItemsLog when card order export has been created
