--- app/models/waste_carriers_engine/card_orders_export_log.rb
@@ -1,31 +1,0 @@
-# frozen_string_literal: true
-
-# This collection tracks card order export files as they are genrerated and accessed
-module WasteCarriersEngine
-  class CardOrdersExportLog
-    include Mongoid::Document
-
-    field :start_time, type: DateTime
-    field :end_time, type: DateTime
-    field :exported_at, type: DateTime
-    field :export_filename, type: String
-    field :first_visited_by, type: User
-    field :first_visited_at, type: DateTime
-
-    def initialize(start_time, end_time, filename, exported_at)
-      super()
-      self.start_time = start_time
-      self.end_time = end_time
-      self.export_filename = filename
-      self.exported_at = exported_at
-    end
-
-    def visit_download_link(user)
-      return if first_visited_by || first_visited_at
-
-      self.first_visited_by = user.id
-      self.first_visited_at = DateTime.now
-      save!
-    end
-  end
-end
